test(refresher): migrate
